Loading mesh Face Normals called inside LoadMeshIndices
This is done because face normals are not depending on the mesh normals (as strange as it sounds). It only depends on the geometry of the mesh, and its builded through its indices. Before this, if a mesh didn't have Normals, it could not have Face Normals, when in reality it should have Face Normals independent from its Normals.